promotions/requirements.txt
pika
python-dotenv
flask
requests
